+++ components/molecules/DashboardScatterChart/dashboard-scatter-chart.tsx
@@ -37,7 +37,8 @@ const DashboardScatterChart: React.FC<DashboardScatterChartProps> = ({ title, op
     <div>
       <div className="flex justify-between px-1">
         <Title level={4} className="!text-sm !font-medium !text-light-slate-12">{title}</Title>
-        <div className="flex flex-col md:flex-row gap-2">
+        {/* replaced display flex to hidden on show/bots container */}
+        <div className="hidden flex-col md:flex-row gap-2">
           <ToggleOption handleToggle={handleShowBots} checked={showBots} optionText="Show Bots"></ToggleOption>
           <ToggleOption
             handleToggle={handleShowMembers}